make library and depdenencies installable
